#에러 해결

+++ src/main/java/com/github/yumyum/map/web/dto/interested/InterestDTO.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class InterestDTO {
         private Integer interestId;
-        private Integer memberId;
+        private Long memberId;
         private Integer rstrId;
         private String rstrName;
         private String address;
